@@ -49,115 +32,53 @@
   "browserslist": [
     "> 1%",
     "last 2 versions",
-    "not ie <= 10"
+    "not dead"
   ],
   "dependencies": {
-    "@ant-design/charts": "^1.2.6",
-    "@ant-design/icons": "^4.0.0",
-    "@ant-design/pro-card": "^1.14.2",
-    "@ant-design/pro-descriptions": "^1.8.3",
-    "@ant-design/pro-form": "^1.29.1",
-    "@ant-design/pro-layout": "^6.24.1",
-    "@ant-design/pro-list": "^1.10.1",
-    "@ant-design/pro-skeleton": "^1.0.1",
-    "@ant-design/pro-table": "^2.42.1",
-    "@types/file-saver": "^2.0.2",
-    "@umijs/hooks": "^1.9.3",
-    "@umijs/route-utils": "^1.0.33",
-    "ahooks": "^2.10.6",
-    "antd": "^4.15.0",
-    "classnames": "^2.2.6",
-    "cnpm": "^6.2.0",
+    "@ant-design/charts": "^2.6.7",
+    "@ant-design/icons": "^5.2.6",
+    "@ant-design/pro-components": "^2.6.0",
+    "@umijs/max": "^4.0.89",
+    "ahooks": "^3.7.8",
+    "antd": "^4.24.15",
+    "classnames": "^2.3.2",
+    "dayjs": "^1.11.10",
     "file-saver": "^2.0.5",
-    "jszip": "^3.6.0",
-    "lodash": "^4.17.11",
-    "moment": "^2.25.3",
+    "jszip": "^3.10.1",
+    "lodash": "^4.17.21",
     "omit.js": "^2.0.2",
-    "rc-resize-observer": "^1.0.0",
-    "react": "^16.14.0",
-    "react-contenteditable": "^3.3.5",
-    "react-dev-inspector": "^1.1.1",
-    "react-dom": "^17.0.0",
-    "react-helmet-async": "^1.0.4",
-    "react-script-tag": "^1.1.2",
-    "react-sortablejs": "^6.0.0",
-    "sortablejs": "^1.14.0",
-    "tyarn": "^0.2.1",
-    "umi": "^3.5.13",
-    "umi-request": "^1.0.8",
-    "universal-cookie": "^4.0.4"
+    "rc-resize-observer": "^1.3.1",
+    "react": "^18.2.0",
+    "react-contenteditable": "^3.3.7",
+    "react-dom": "^18.2.0",
+    "react-helmet-async": "^1.3.0",
+    "react-sortablejs": "^6.1.4",
+    "sortablejs": "^1.15.0",
+    "universal-cookie": "^6.1.1"
   },
   "devDependencies": {
-    "@ant-design/pro-cli": "^1.0.28",
-    "@types/classnames": "^2.2.7",
-    "@types/express": "^4.17.0",
-    "@types/history": "^4.7.2",
-    "@types/jest": "^26.0.0",
-    "@types/lodash": "^4.14.144",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-helmet": "^6.1.0",
-    "@types/sortablejs": "^1.10.7",
-    "@umijs/fabric": "^2.5.1",
-    "@umijs/plugin-blocks": "^2.0.5",
-    "@umijs/plugin-esbuild": "^1.0.1",
-    "@umijs/preset-ant-design-pro": "^1.2.0",
-    "@umijs/preset-react": "^1.4.8",
-    "@umijs/yorkie": "^2.0.3",
-    "carlo": "^0.9.46",
-    "chalk": "^4.0.0",
-    "cross-env": "^7.0.0",
-    "cross-port-killer": "^1.1.1",
-    "detect-installer": "^1.0.1",
-    "enzyme": "^3.11.0",
-    "eslint": "^7.1.0",
-    "express": "^4.17.1",
-    "gh-pages": "^3.0.0",
-    "jsdom-global": "^3.0.2",
-    "lint-staged": "^10.0.0",
-    "mockjs": "^1.0.1-beta3",
-    "prettier": "^2.0.1",
-    "puppeteer-core": "^8.0.0",
-    "stylelint": "^13.0.0",
-    "typescript": "^4.0.3"
+    "@types/classnames": "^2.3.1",
+    "@types/file-saver": "^2.0.5",
+    "@types/lodash": "^4.14.202",
+    "@types/react": "^18.2.42",
+    "@types/react-dom": "^18.2.17",
+    "@types/sortablejs": "^1.15.7",
+    "@typescript-eslint/eslint-plugin": "^6.14.0",
+    "@typescript-eslint/parser": "^6.14.0",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.55.0",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "husky": "^8.0.3",
+    "lint-staged": "^15.2.0",
+    "postcss-less": "^6.0.0",
+    "prettier": "^3.1.1",
+    "stylelint": "^15.11.0",
+    "stylelint-config-standard": "^34.0.0",
+    "stylelint-order": "^6.0.4",
+    "typescript": "^5.3.3"
   },
   "engines": {
-    "node": ">=10.0.0"
-  },
-  "checkFiles": [
-    "src/**/*.js*",
-    "src/**/*.ts*",
-    "src/**/*.less",
-    "config/**/*.js*",
-    "scripts/**/*.js"
-  ],
-  "create-umi": {
-    "ignoreScript": [
-      "docker*",
-      "functions*",
-      "site",
-      "generateMock"
-    ],
-    "ignoreDependencies": [
-      "netlify*",
-      "serverless"
-    ],
-    "ignore": [
-      ".dockerignore",
-      ".git",
-      ".github",
-      ".gitpod.yml",
-      "CODE_OF_CONDUCT.md",
-      "Dockerfile",
-      "Dockerfile.*",
-      "lambda",
-      "LICENSE",
-      "netlify.toml",
-      "README.*.md",
-      "azure-pipelines.yml",
-      "docker",
-      "CNAME",
-      "create-umi"
-    ]
+    "node": ">=18.0.0"
   }
 }
