@@ -1,7 +1,7 @@
-import React from 'react';
-import {PageLoading} from '@ant-design/pro-layout';
-import type {ConnectProps} from 'umi';
-import {connect, Redirect} from 'umi';
+import React, {useEffect, useState} from 'react';
+import {PageLoading} from '@ant-design/pro-components';
+import type {ConnectProps} from '@umijs/max';
+import {connect, Navigate, Outlet} from '@umijs/max';
 import {stringify} from 'querystring';
 import type {ConnectState} from '@/models/connect';
 import type {StaffAdminInterface} from '@/services/staffAdmin';
@@ -10,52 +10,36 @@ import {LSExtStaffAdminID} from '../../config/constant';
 type SecurityLayoutProps = {
   loading?: boolean;
   currentStaffAdmin?: StaffAdminInterface;
-} & ConnectProps;
-
-type SecurityLayoutState = {
-  isReady: boolean;
+  dispatch?: ConnectProps['dispatch'];
 };
 
-class StaffAdminSecurityLayout extends React.Component<SecurityLayoutProps, SecurityLayoutState> {
-  state: SecurityLayoutState = {
-    isReady: false,
-  };
+const StaffAdminSecurityLayout: React.FC<SecurityLayoutProps> = (props) => {
+  const {dispatch, loading} = props;
+  const [isReady, setIsReady] = useState(false);
 
-  componentDidMount() {
-    this.setState({
-      isReady: true,
-    });
-    const {dispatch} = this.props;
+  useEffect(() => {
+    setIsReady(true);
     if (dispatch) {
-      // dispatch({
-      //   type: 'adminType/changeStatus',
-      //   payload: 'staffAdmin',
-      // });
       dispatch({
         type: 'staffAdmin/getCurrent',
       });
     }
-  }
+  }, [dispatch]);
 
-  render() {
-    const {isReady} = this.state;
-    const {children, loading} = this.props;
-    // You can replace it to your authentication rule (such as check token exists)
-    // You can replace it with your own login authentication rules (such as judging whether the token exists)
-    const isLogin = localStorage.getItem(LSExtStaffAdminID) !== null;
-    const queryString = stringify({
-      redirect: window.location.href,
-    });
+  // You can replace it with your own login authentication rules (such as judging whether the token exists)
+  const isLogin = localStorage.getItem(LSExtStaffAdminID) !== null;
+  const queryString = stringify({
+    redirect: window.location.href,
+  });
 
-    if ((!isLogin && loading) || !isReady) {
-      return <PageLoading/>;
-    }
-    if (!isLogin && window.location.pathname !== '/staff-admin/login') {
-      return <Redirect to={`/staff-admin/login?${queryString}`}/>;
-    }
-    return children;
+  if ((!isLogin && loading) || !isReady) {
+    return <PageLoading/>;
   }
-}
+  if (!isLogin && window.location.pathname !== '/staff-admin/login') {
+    return <Navigate to={`/staff-admin/login?${queryString}`} replace />;
+  }
+  return <Outlet />;
+};
 
 export default connect(({staffAdmin, loading}: ConnectState) => ({
   currentStaffAdmin: staffAdmin.currentStaffAdmin,
