@@ -1,57 +1,66 @@
 // https://umijs.org/config/
-import {defineConfig} from 'umi';
+import { defineConfig } from '@umijs/max';
 import proxy from './proxy';
 import routes from './routes';
 import theme from './theme';
 
-const {REACT_APP_ENV} = process.env;
+const { REACT_APP_ENV } = process.env;
 
 export default defineConfig({
   hash: true,
-  antd: {},
-  dva: {
-    hmr: true,
-  },
   history: {
     type: 'browser',
   },
-  dynamicImport: {
-    loading: '@/components/PageLoading/index',
-  },
-  targets: {
-    ie: 11,
-  },
-  // umi routes: https://umijs.org/docs/routing
   routes,
-  // https://umijs.org/zh-CN/config#theme
-  theme: theme,
   title: false,
-  ignoreMomentLocale: true,
-  proxy: proxy[REACT_APP_ENV || 'dev'],
-  manifest: {
-    basePath: '/',
+
+  // 定义环境变量供客户端使用
+  define: {
+    'process.env.REACT_APP_ENV': REACT_APP_ENV,
   },
-  // 快速刷新功能 https://umijs.org/config#fastrefresh
-  fastRefresh: {},
-  esbuild: {},
-  // webpack5: {},
+
+  // antd 配置
+  antd: {
+    configProvider: {},
+    // 使用 dayjs 替代 moment
+    momentPicker: false,
+  },
+
+  // dva 数据流
+  dva: {},
+
+  // 主题配置
+  theme: theme,
+
+  // 代理配置
+  proxy: proxy[REACT_APP_ENV || 'dev'],
+
+  // 请求配置
   request: {
     dataField: 'data',
   },
-  scripts: [
-    'https://unpkg.com/react@17/umd/react.production.min.js',
-    'https://unpkg.com/react-dom@17/umd/react-dom.production.min.js',
-    'https://unpkg.com/@ant-design/charts@1.0.5/dist/charts.min.js',
-    //使用 组织架构图、流程图、资金流向图、缩进树图 才需要使用
-    //'https://unpkg.com/@ant-design/charts@1.0.5/dist/charts_g6.min.js',
-  ],
-  externals: {
-    react: 'React',
-    'react-dom': 'ReactDOM',
-    "@ant-design/charts": "charts"
+
+  // 外部脚本 - 仅在生产环境使用 CDN
+  // headScripts: [
+  //   'https://unpkg.com/react@18/umd/react.production.min.js',
+  //   'https://unpkg.com/react-dom@18/umd/react-dom.production.min.js',
+  //   'https://unpkg.com/@ant-design/charts@1.4.0/dist/charts.min.js',
+  // ],
+
+  // externals 配置 - 仅在生产环境使用
+  // externals: {
+  //   react: 'React',
+  //   'react-dom': 'ReactDOM',
+  //   '@ant-design/charts': 'charts',
+  // },
+
+  // manifest
+  manifest: {
+    basePath: '/',
   },
-  devServer: {
-    host: 'localhost',
-    port: 9000,
+
+  // 代码分割
+  codeSplitting: {
+    jsStrategy: 'granularChunks',
   },
 });
