@@ -1,4 +1,4 @@
-import { PageLoading } from '@ant-design/pro-layout';
+import { PageLoading } from '@ant-design/pro-components';
 
 // loading components from code split
 // https://umijs.org/plugin/umi-plugin-react.html#dynamicimport
