@@ -1,6 +1,6 @@
 import { Button, Result } from 'antd';
 import React from 'react';
-import { history } from 'umi';
+import { history } from '@umijs/max';
 
 const NoFoundPage: React.FC = () => (
   <Result
