@@ -1,15 +1,13 @@
 import React, {useRef, useState} from 'react';
-import {FooterToolbar, PageContainer} from '@ant-design/pro-layout';
+import type {ActionType, ProColumns} from '@ant-design/pro-components';
+import {FooterToolbar, PageContainer, ProTable} from '@ant-design/pro-components';
 import {Badge, Button, Divider, Modal} from 'antd';
 import {NotificationOutlined, PlusOutlined} from '@ant-design/icons';
 import type {CustomerMassMsgItem} from '@/pages/StaffAdmin/CustomerMassMsg/data';
 import {Delete, Notify, Query} from '@/pages/StaffAdmin/CustomerMassMsg/service';
-import type {ProColumns} from '@ant-design/pro-table/es';
 import {history} from '@@/core/history';
-import type {ActionType} from '@ant-design/pro-table';
-import ProTable from '@ant-design/pro-table';
 import {HandleRequest, ProTableRequestAdapter} from '@/utils/utils';
-import moment from 'moment';
+import dayjs from 'dayjs';
 import type {WelcomeMsg} from '@/pages/StaffAdmin/CustomerWelcomeMsg/data';
 import AutoReplyPreviewModal from '@/pages/StaffAdmin/Components/Modals/AutoReplyPreviewModal';
 
@@ -64,7 +62,7 @@ const CustomerMassMsgList: React.FC = () => {
         return (
           <div
             dangerouslySetInnerHTML={{
-              __html: moment(item.created_at).format('YYYY-MM-DD HH:mm').split(' ').join('<br />'),
+              __html: dayjs(item.created_at).format('YYYY-MM-DD HH:mm').split(' ').join('<br />'),
             }}
           />
         );
