@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { history } from '@@/core/history';
-import { connect } from 'umi';
+import { connect } from '@umijs/max';
 import type { Dispatch } from '@@/plugin-dva/connect';
 
 export type LoginCallbackProps = {
