@@ -1,12 +1,12 @@
 import { LeftOutlined } from '@ant-design/icons';
-import { PageContainer } from '@ant-design/pro-layout';
-import { history } from 'umi';
+import { PageContainer } from '@ant-design/pro-components';
+import { history } from '@umijs/max';
 
 const MemberDetailInfo = (props: any) => {
     console.log(props.location.query)
     return (
         <PageContainer
-            onBack={() => history.goBack()}
+            onBack={() => history.back()}
             backIcon={<LeftOutlined />}
             header={{
                 title: '成员详情',
